organize folder structure and indoc module name
